Add view fixture to toplevel conftest
