+++ gaphor/UML/classes/tests/test_classeseditors.py
@@ -1,6 +1,5 @@
 # ruff: noqa: F401,F811
 from gaphor import UML
-from gaphor.diagram.tests.test_instanteditors import view
 from gaphor.UML import recipes
 from gaphor.UML.classes.classeseditors import association_item_editor
