+++ gaphor/diagram/general/tests/test_generaleditors.py
@@ -2,7 +2,6 @@
 from gaphor.core.modeling import Comment
 from gaphor.diagram.general import CommentItem
 from gaphor.diagram.general.generaleditors import comment_item_editor
-from gaphor.diagram.tests.test_instanteditors import view
 
 
 def test_comment_item_editor(diagram, element_factory, view, event_manager):
