+++ gaphor/diagram/tests/test_instanteditors.py
@@ -1,20 +1,5 @@
-import pytest
-from gaphas.view import GtkView
-
 from gaphor import UML
 from gaphor.diagram.instanteditors import named_item_editor
-from gaphor.diagram.painter import ItemPainter
-from gaphor.diagram.selection import Selection
-
-
-@pytest.fixture
-def view(diagram):
-    view = GtkView(model=diagram, selection=Selection())
-    view._qtree.resize((-100, -100, 400, 400))
-    item_painter = ItemPainter(view.selection)
-    view.painter = item_painter
-    view.bounding_box_painter = item_painter
-    return view
 
 
 def test_named_item_editor_with_element(diagram, element_factory, view, event_manager):
